Committing the first version of the encounter.jsp file -- now with AJAX!
In order to be explicit, I tried to comment all of the places in the
encounter.jsp file where I added the ajax code, as well as detailed why
the edits were necessary. 

This commit consists of three files:
encounter.jsp (edits)
ajaxSubmit.js (new)
EncounterSetReleaseData.java (removes the encounter link from the error)

The ajaxSubmit.js is the main driver for our updates to the
encounter.jsp page. It works by serializing the current forms and using
a jquery post to send the data to the already existing handlers.