AJAX updates : Metadata Section
This commit updates the Metadata section, allowing most of the pieces to
use the new ajax code. 

Notes: 

The Managing Researcher section still, for the most part, uses the old
form submission for the following Reason. When the managing researcher
is set to someone other than the current user, the current user loses
edit access to the encounter. Thusly, when this change happens we need
to reload the page to show the current user that they no longer have
access.

Delete Encounter for the most part uses the ajax implementation to check
and make sure that the delete was successful. If it fails, the error is
displayed, and if it succeeds we are redirected to the unapproved
encounters page.